[Kernel][IcebergWriterCompatV1] Update column mapping so physicalName='col-{columnId}' and enforce this as part of our checks (#4383)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-io/delta/blob/master/CONTRIBUTING.md
2. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  3. Be sure to keep the PR description updated to reflect all changes.
  4. Please write your PR title to summarize what this PR proposes.
5. If possible, provide a concise example to reproduce the issue for a
faster review.
6. If applicable, include the corresponding issue number in the PR title
and link it in the body.
-->

#### Which Delta project/connector is this regarding?
<!--
Please add the component selected below to the beginning of the pull
request title
For example: [Spark] Title of my pull request
-->

- [ ] Spark
- [ ] Standalone
- [ ] Flink
- [x] Kernel
- [ ] Other (fill in here)

## Description

Part of https://github.com/delta-io/delta/issues/4289

Per-the-RFC when icebergWriterCompatV1 is enabled we want column mapping
physicalName to be equal to `col-[fieldId]`. This PR enforces that as
part of our icebergWriterCompatV1 metadata validations, and also updates
our column mapping code to follow that behavior when
icebergWriterCompatV1 is enabled.

## How was this patch tested?

Adds new tests and modifies existing tests to check the new changes.

---------

Co-authored-by: Nick Lanham <nicklan@users.noreply.github.com>