[Spark] Add option to DeltaParquetFileFormat to allow missing Row Tracking metadata fields (#4387)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-io/delta/blob/master/CONTRIBUTING.md
2. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  3. Be sure to keep the PR description updated to reflect all changes.
  4. Please write your PR title to summarize what this PR proposes.
5. If possible, provide a concise example to reproduce the issue for a
faster review.
6. If applicable, include the corresponding issue number in the PR title
and link it in the body.
-->

#### Which Delta project/connector is this regarding?
<!--
Please add the component selected below to the beginning of the pull
request title
For example: [Spark] Title of my pull request
-->

- [x] Spark
- [ ] Standalone
- [ ] Flink
- [ ] Kernel
- [ ] Other (fill in here)

## Description

This PR:

- Renames `nullableRowTrackingFields` in `DeltaParquetFileFormat` to
`nullableRowTrackingGeneratedFields`.
- Adds another field called `nullableRowTrackingConstantFields` in
`DeltaParquetFileFormat`, which allows reading row ids and row commit
versions when the `baseRowId` and `defaultRowCommitVersion`s are missing
(returning a NULL in this case).
- It adds a helper function `transformFileFormat` to `DeltaTableUtils`
to make it easier to set the new field.

## How was this patch tested?

Added a test to `RowIdSuite`.

## Does this PR introduce _any_ user-facing changes?

No