[Kernel] Read the fileSizeHistogram from CRCInfo and propagate it through transaction metrics (#4403)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-io/delta/blob/master/CONTRIBUTING.md
2. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  3. Be sure to keep the PR description updated to reflect all changes.
  4. Please write your PR title to summarize what this PR proposes.
5. If possible, provide a concise example to reproduce the issue for a
faster review.
6. If applicable, include the corresponding issue number in the PR title
and link it in the body.
-->

#### Which Delta project/connector is this regarding?
<!--
Please add the component selected below to the beginning of the pull
request title
For example: [Spark] Title of my pull request
-->

- [ ] Spark
- [ ] Standalone
- [ ] Flink
- [x] Kernel
- [ ] Other (fill in here)

## Description

Reads the fileSizeHistogram from CRC and add it as part of transaction
metrics. Updates it as the transaction is executed.

Note, this is a little weird since it only adds read support and no
write support, so we only actually have histograms available for new
tables in E2E tests. But https://github.com/delta-io/delta/pull/4328
will add write support and fill out the full feature afterward.
Documented throughout the changes needed when adding write support.

## How was this patch tested?

Updates tests.

## Does this PR introduce _any_ user-facing changes?

No